binary tree
Create a binary tree for each of the languages

The Ruby and JavaScript implementations define a `BinaryTree`
class/prototype function. `insert` and `delete` are destructive.
`delete` returns the value or `nil`/`null` if it doesn't
exist. Insert returns the updated tree.

For c, there are two implementations: a binary tree and a threaded
binary tree. The regular tree uses recursion and the threaded binary
tree uses only iteration. (The threaded tree is only threaded for
traversal then it's de-threaded.) `delete` returns a `bool`.

For Scheme, the binary tree is persistent. Insert and delete return a
tree.